+++ crates/cli/src/commands/plugin/add.rs
@@ -1,7 +1,7 @@
 use crate::session::ProtoSession;
 use clap::Args;
 use iocraft::prelude::element;
-use proto_core::{Id, PinLocation, PluginLocator, ProtoConfig, cfg};
+use proto_core::{Id, PinLocation, PluginLocator, PluginType, ProtoConfig, cfg};
 use starbase::AppResult;
 use starbase_console::ui::*;
 
@@ -15,18 +15,37 @@ pub struct AddPluginArgs {
 
     #[arg(long, default_value_t, help = "Location of .prototools to add to")]
     to: PinLocation,
+
+    #[arg(long = "type", default_value_t, help = "The type of plugin to add")]
+    ty: PluginType,
 }
 
 #[tracing::instrument(skip_all)]
 pub async fn add(session: ProtoSession, args: AddPluginArgs) -> AppResult {
     let config_path = ProtoConfig::update_document(session.env.get_config_dir(args.to), |doc| {
-        let plugins = doc["plugins"].or_insert(cfg::table());
-        plugins[args.id.as_str()] = cfg::value(args.plugin.to_string());
+        let key = if args.ty == PluginType::Backend {
+            "backends"
+        } else {
+            "tools"
+        };
 
-        // config
-        //     .plugins
-        //     .get_or_insert_default()
-        //     .insert(args.id.clone(), args.plugin.clone());
+        // Convert legacy [plugins] to [plugins.tools]
+        if doc["plugins"]
+            .as_table()
+            .is_some_and(|table| !table.contains_key("backends") && !table.contains_key("tools"))
+        {
+            let existing = doc["plugins"].clone();
+
+            doc.remove("plugins");
+
+            let plugins = doc["plugins"].or_insert(cfg::implicit_table());
+            plugins["tools"] = existing;
+        }
+
+        // Add plugin to nested tables
+        let plugins = doc["plugins"].or_insert(cfg::implicit_table());
+        let table = plugins[key].or_insert(cfg::table());
+        table[args.id.as_str()] = cfg::value(args.plugin.to_string());
     })?;
 
     // Load the tool and verify it works. We can't load the tool with the